requirements.txt
requests
python-telegram-bot
python-dotenv
pyyaml
python-binance
typing